+++ tool/build.dart
@@ -1,6 +1,0 @@
-import 'dart:io';
-
-void main() {
-  var uri = Platform.script;
-  print(uri.toFilePath());
-}